[Angular] Theming changes and project organization (#64)
* Add theme configuration to Angular example

Sets up the theme for the Angular restaurant example.

* Add theme bindings to all components

Updates all of the components to include the theme class/style bindings.

* Set up Angular catalog as a library

Sets up the catalog of Angular components as a separate library so that it's easier to reuse across demos.

* Set up restaurant demo as a separate project

Sets up the restaurant demo as a separate Angular project so that it's easier to spin up more demos.

* Add common function for configuring A2UI

Adds the `provideA2UI` function so that we have better type safety when setting up a new demo.